coll: remove errflag from MPIC_Wait and MPIC_Recv
It makes sense that we want to propage errflag in send via tag bits and
receive that errflag. However, it doesn't make sense to compress the
receive error into errflag, resulting in loss of error information.

The errorbit in the tag should be checked in
MPIR_Request_completion_processing. Then it no longer need be returned
in MPIC_Wait or MPIC_Recv. In fact, most of the usage simply ignores the
returned errflag anyway.