datatype: avoid divide by 0 in get_elements_x
When the datatype is a struct with mixed elements, we set its basic_type
to MPI_DATATYPE_NULL. This certainly is not ideal. But for now, lets
avoid divide by 0 error in MPIR_Type_get_basic_type_elements and simply
return 0.

MPIR_Type_get_basic_type_elements is used in the OFI receive complete
event to check whether we received partial elements. We'll skip the
error checking for struct types for now.