comm: avoid creating invalid intercomm then destroy it
In MPIR_Comm_create_inter, we know whether the remote group is empty
after the exchange, thus it is unnecessary to create and commit the
intercomm then delete it later. Simply don't create it in the first
place.

The device layer is not necessarily equipped to handle intercomm commit
with empty groups.